+++ eLibrary/UI/elibrary_mobile/lib/models/uplata.dart
@@ -1,3 +1,5 @@
+import 'dart:ffi';
+
 import 'package:elibrary_mobile/models/biblioteka.dart';
 import 'package:elibrary_mobile/models/citalac.dart';
 import 'package:elibrary_mobile/models/tip_uplate.dart';
@@ -10,7 +12,7 @@ class Uplata {
   int? uplataId;
   int? citalacId;
   int? bibliotekaId;
-  int? iznos;
+  double? iznos;
   String? datumUplate;
   int? tipUplateId;
   int? valutaId;
